requirements.txt
<<<<<<< HEAD
flask>=3.0.0
rembg[cpu]>=2.0.68
Pillow>=10.2.0
Werkzeug>=3.0.1
gunicorn
onnxruntime
numpy
=======
flask>=3.0.0
rembg[cpu]>=2.0.68
Pillow>=10.2.0
Werkzeug>=3.0.1
gunicorn
onnxruntime
numpy
>>>>>>> 629bba4986429f2b38923f3a1fe933ccf0a1c15e
